+++ src/di/auth_repository.py
@@ -1,0 +1,3 @@
+from src.repository.auth_repository import AuthRepository
+
+auth_repository = AuthRepository()
