+++ src/repository/password_repository.py
@@ -1,17 +1,16 @@
 import os
-from datetime import timedelta, datetime
+import shutil
 
-import jwt
-from fastapi import HTTPException
-from passlib.context import CryptContext
+from fastapi import HTTPException, UploadFile, File
 from sqlalchemy import func
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm.exc import NoResultFound
+from starlette.responses import FileResponse
 
 from src.config.db_connection import session, engine
+from src.di.auth_repository import auth_repository
 from src.entity.credential import Credential, Base
 from src.dto.credential_response import CredentialResponse
-from src.entity.token import TokenBlacklist
 from src.entity.user import User
 from src.dto.user_response import UserResponse
 
@@ -19,12 +18,6 @@ Base.metadata.create_all(bind=engine)
 
 
 class PasswordRepository:
-    pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-    authenticated_user = None
-
-    secret_key = os.getenv("SECRET_KEY")
-    algorithm_type = os.getenv("ALGORITHM")
 
     def create_password(self, password_data: CredentialResponse, user_id):
         data = Credential(
@@ -76,7 +69,7 @@ class PasswordRepository:
                 user.username = new_user_data.username
                 user.email = new_user_data.email
                 if new_user_data.password:
-                    user.password = self.get_password_hash(new_user_data.password)
+                    user.password = auth_repository.get_password_hash(new_user_data.password)
                 session.commit()
                 return {"Message": "User updated successfully"}
             else:
@@ -85,68 +78,29 @@ class PasswordRepository:
             session.rollback()
             raise HTTPException(status_code=400, detail="Email already exists")
 
-    def get_password_hash(self, password):
-        return self.pwd_context.hash(password)
+    def upload_file(self, file: UploadFile = File(...)):
+        destination_folder = os.getenv("DESTINATION_FOLDER")
+
+        # Move the uploaded file to the destination folder
+        file_path = destination_folder + "/" + file.filename
+        with open(file_path, "wb") as buffer:
+            shutil.copyfileobj(file.file, buffer)
+
+        return {"filename": file.filename, "destination_folder": destination_folder}
+
+    def read_data(self):
+        file_path = os.getenv("FILE_PATH")
+        response = FileResponse(file_path, media_type="text/csv")
+        response.headers["Content-Disposition"] = "attachment; filename=example.csv"
+        return response
 
     def sign_up(self, new_user: UserResponse):
         user = User(username=new_user.username,
                     email=new_user.email,
-                    password=self.get_password_hash(new_user.password))
+                    password=auth_repository.get_password_hash(new_user.password))
         try:
             session.add(user)
             session.commit()
             return {"Message": "New Account Created"}
         except HTTPException:
             raise HTTPException(status_code=400, detail="something went wrong")
-
-    def create_access_token(self, data: dict, expires_delta: timedelta):
-        to_encode = data.copy()
-        expire = datetime.now() + expires_delta
-
-        to_encode.update({"exp": expire})
-        encoded_jwt = jwt.encode(to_encode, self.secret_key, algorithm=self.algorithm_type)
-
-        # db_token_exp = TokenExpiration(token=encoded_jwt, expire_at=expire)
-        # session.add(db_token_exp)
-        # session.commit()
-
-        return encoded_jwt
-
-    def revoke_token(self, token: str):
-        # Store revoked token in the database
-        db_token_blacklist = TokenBlacklist(token=token)
-        session.add(db_token_blacklist)
-        session.commit()
-
-    # def is_token_expired(self, token: str) -> bool:
-    #     # Check if token is expired
-    #     token_exp = session.query(TokenExpiration).filter_by(token=token).first()
-    #     return token_exp.expire_at < datetime.now() if token_exp else True
-
-    def is_token_revoked(self, token: str) -> bool:
-        # Check if token is revoked
-        return session.query(TokenBlacklist).filter_by(token=token).first() is not None
-
-    def retrieve_user_id(self):
-        if self.authenticated_user is None:
-            raise HTTPException(status_code=400, detail="User not authenticated")
-        return self.authenticated_user.id
-
-    def authenticate_user(self, username, password):
-        try:
-            user = session.query(User).filter_by(username=username).one()
-            if not user:
-                raise HTTPException(status_code=400, detail="User not found")
-            if not self.pwd_context.verify(password, user.password):
-                raise HTTPException(status_code=400, detail="Password incorrect")
-            self.authenticated_user = user
-            return self.token_access(user)
-        except NoResultFound:
-            raise HTTPException(status_code=400, detail="User not found")
-        finally:
-            session.close()
-
-    def token_access(self, user):
-        access_token = self.create_access_token(data={"sub": user.username, "user_id": user.id},
-                                                expires_delta=timedelta(minutes=30))
-        return {"user_id": user.id, "access_token": access_token, "token_type": "bearer"}
